Move numpy test to test_optional
