+++ packages/python/plotly/plotly/tests/test_core/test_subplots/test_make_subplots.py
@@ -1,7 +1,5 @@
 from __future__ import absolute_import
 
-import numpy as np
-
 from unittest import TestCase
 from plotly.graph_objs import (
     Annotation,
@@ -1559,17 +1557,6 @@ class TestMakeSubplots(TestCase):
         expected = tls.make_subplots(insets=[{"cell": (1, 1), "l": 0.7, "b": 0.3}])
         fig = tls.make_subplots(
             insets=[{"cell": (1, 1), "l": 0.7, "b": 0.3}], subplot_titles=[]
-        )
-        self.assertEqual(fig, expected)
-
-    def test_subplot_titles_numpy_array(self):
-        # Pass numpy array
-        expected = tls.make_subplots(
-            insets=[{"cell": (1, 1), "l": 0.7, "b": 0.3}], subplot_titles=("", "Inset")
-        )
-        fig = tls.make_subplots(
-            insets=[{"cell": (1, 1), "l": 0.7, "b": 0.3}],
-            subplot_titles=np.array(["", "Inset"]),
         )
         self.assertEqual(fig, expected)
